Fix #90
