+++ src/PInvoke.sln
@@ -124,7 +124,6 @@ Global
 		Kernel32.Tests.Shared\Kernel32.Tests.Shared.projitems*{d7224f5b-1aee-4624-a71f-f0bb10c63555}*SharedItemsImports = 4
 		Kernel32.Shared\Kernel32.Shared.projitems*{18ba8c60-9a63-4eea-bb5c-c9899d86f5b8}*SharedItemsImports = 4
 		AdvApi32.Shared\AdvApi32.Shared.projitems*{46eca0bd-f083-481b-a519-c4c910432f65}*SharedItemsImports = 4
-		Kernel32.Shared\Kernel32.Shared.projitems*{46eca0bd-f083-481b-a519-c4c910432f65}*SharedItemsImports = 4
 	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
